+++ components/armDecoder/SemanticActions.hpp
@@ -108,7 +108,7 @@ struct Operation {
   virtual ~Operation() {}
   virtual Operand operator()( std::vector< Operand > const & operands ) = 0;
   virtual Operand evalExtra( std::vector< Operand > const & operands ) {
-    return 0;
+    return (uint64_t)0;
   }
   virtual char const * describe() const = 0;
